Refactor imports and self-extension checks
- Make local imports explicit for emitted ESM modules
- Centralize recursive self-extension detection across layouts
